begin switchover to peripherals
A few commits ago I added a design document for the scheduler
which includes switching from sequences to a more abstract idea of
peripherals. This commit introduces utils/peripherals.js, which
helps to work with peripherals, and it duplicates the scheduling
tests that already existed to verify that peripherals include all
th data that sequences included. The next step will be to rewrite
the scheduler which is now gpio/gpio.js, and to add tests for all
the schedule types noted in the scheduler design doc.