Refactored UI
The UI is still extremely weak, but this introduces several
improvements.

The UI is more closely aligned with the current peripheral-based
schema.

I deleted some unused code and split the UI code into multiple files.
There are now /component, /form, and /datavis subdirectories under
client for better code organization.

I added the ability to create sequenceItems and Peripherals through
the UI. The peripherals especially leave much to be desired, because
you can't yet specify a peripheral type (they are automatically
'RELAY') or a sequence type (automatically 'DURATION').

I also added a demo-adapter to the water-shape api adapters. This is a
simple adapter that lets you use a simple object in the browser as a
data store, making it possible to demo all the functionality of the app
using a static site. I added compile-demo and watch-demo tasks to
compile the site with the demo adapter instead of the regular one, but
I did not create a task to serve the demo site locally because it
looked like a more involved change and I want to think about how to do
configuration in a holistic way.

The immediate use for the demo-adapter is to create a gh-pages demo of
the app.