added functions for time periods to timeParser
These functions are required for adding validation to sequenceItems
(mostly done) and for comparing times.

Also updated editable to use save rather than update to work around a
bug in validation--when I save an incomplete object the client-side
validation would have to do a GET to get the whole object, which would
probably be more overhead than is saved by doing a PUT instead of POST.

There's another validation bug that when the back end returns an error
to the fromt end as a regular string, JSON parsing breaks somewhere and
I don't get the error I want.

Updated sequenceManipulation to use timeParser but that whole file
should be on its way out.

Still need to update all the validation login in water-shape to enforce
a timeout on validators after which they are assumed to have failed, and
also make sure they are safe if callbacks are called more than once.
Need tests for those cases.