fix bug in time parser
The regex for times used to accept 15:560, and the timeparser
interpreted it as 15:56:00. I made the timeparser more selective to
reject that.

I had previously included a step where if my regex for the time didn't
match, I would try to parse it by doing `new Date(t)`. However while
making this change I found that `new Date('15:560') does not fail, and
is interpreted as 'Tue Jan 01  560 15:00:00 GMT-0500'. Now that I have
an example of a case where doing `new Date(t)` leads to incorrect
behavior, I am no longer using it if the time regexes fail.