generic tables return the 0th element of 'single' arrays
When a constructed table entry specified 'single', water-shape
used to return the 0th element in the found array UNLESS there was
not exactly one element. Now it always returns the 0th element, with
the side effect that when the list is empty it returns [][0], which is
`undefined`.

Also added test.