Updated request adapter to parse own responses
Setting json:true in request args causes request to throw an error
if the response is not json, but I wanted to have the option to just
use a non-json string. So I updated the DMI methods in request-adapter
to try to parse out json but simply return the body of the response if
that fails. The REST methods--.get, .put, .post--still specify
json:true. .delete does not because the response is usually undefined.