update editablevalue to allow for no displayvalue
The initially displayed value is an ugly part of the editablevalue api.

When I made it, I imagined that the editableValue might maintain the
state of the value to display itself. In fact, editableValues just get
re-rendered when their values change, so the value to display should be
as basic a value as possible, like string, boolean, or number.