Add delete button for sequence item and peripheral
Added peripheral delete in sequenceManipulation.js. Currently
it deletes the peripheral rule and peripheral override rules with the
peripheral as test or subject. It leaves the sequence and sequenceItems.
I keep building based on this idea that people will want to save
sequences but I'm getting the feeling that's dumb. And it's not like I
can't implement cloning whole peripherals later, with all their
subrecords.

Anyway, this is how it is now.