Update hapi adapter to try to deserialize json
Aparently reqest, which I use to test the api requestAdapter,
and browser-request, which I use in the front end, handle their
arguments slightly differently. When you don't specify json:true
in browser-request it uses the text/plain MIME type, which the server
interprets as a string rather than json. Apparently server-side request
uses something else, which the server interprets as json correctly.

I fixed the immediate problem by having the server try to deserialize
request payloads for PUTs and POSTs. But the bigger problem is that I'm
not actually unit testing the request adapter in water-shape the same
way I'm using it. browser-request doesn't work on the server, so I'm
going to have to find a way to run the tests from inside the browser.