Added tests for request-adapter using browser-request
I had to add phantom into the unit tests to get this to work. The issue
was that browser-request does not run in node--it can only run in a
browser. So to run the tests, I created a new server configuration that
serves a page with the browser-request-backed api object in window
scope, then use a wrapper comprised of webdriver.executeAsyncScript to
make it look like a node-native dmi object to the tests. It's pretty
hacky but it still feels a lot better than not actually testing
browser-request at all.