use in-memory db to avoid timing issues in test
The scheduling tests kept failing in Travis because of what seemed like
a fencepost error. I changed the test so that the timing isn't a factor
of the scheduler reload time. I also changed the scheduler so that it
will not write 0 to the gpio pin if the current state is already 0.

I also noticed that the tests were duplicated in two files, so I
deleted one of them.