Add support for literals in tree tables
I needed to have a field in a constructed table that selected all the
gpioPin records with a peripheralId of null. The existing generic tree
table implementation, which was called a constructed table, didn't
allow that.

I refactored it in two ways. First, I changed the name from constructed
table to tree. I also changed the schema so that instead of identifying
constructed tables by using `constructed: true` in the table structure,
every table has a `type` attribute. The default type attribute is
`PERSISTED`, which corresponds to a table actually saved in the
database. What used to be a constructed table is now a table of type
`TREE`. TREE tables have an attribute called `root`. This is the
name of the table that is listed when the `list` method is called, and
which is searched when the `getById` method is called. a TREE table can
only have one root, and must have a root.

The switch from `constructed: true` to specifying a `type` attribute
opens the door for further types, such as a table that represents every
pair of records from two tables that satisfy some function. I don't
have specific plans to implement any other table types, but now that
I've had to implement one, I'm making it possible to implement more.