Add editable fields for gpio peripheral type dependencies
The peripheral component now reads the gpio dependencies for its peripheral
type, displays the gpios that are assigned to those dependencies, or
'None', if no gpio is assigned, and allows the user to select an unused
gpio.

This includes an update of the delete behavior for peripherals. Instead
of the incorrect behavior of trying to delete gpio records, it updates
any associated gpios, setting their peripheralId,
peripheralTypeDependency, and ioType to null.

It also includes an update of the editable to clarify how client code
should provide a function to be called on a successful update. Clients
can now include an onUpdate attribute on the options object, which will
be called on a successful update.