Initial notes on a scheduler
The scheduler reads the completePeripherals from the dmi.

It finds the dependencies, populates them, and uses them to instantiate
a controller.

If the sequenceType is DURATION, the controller runs on its own, cycling
through the sequenceItems. If TIME, it registers with the tick callback.

Still haven't sorted out reconciling schedule changes, dealing with
exceptions when trying to create controllers, tests, and triggered
peripherals.