Added alignment to sequences.
The alignment field is the time of dat when a duration sequence starts.
This means that whenever the scheduler starts up, the duration
sequences will be treated as if they had started cycling at the time
specified in their alignment. This will make it easier to update
running sequences and for users to know what to expect. Alignment is
not a required field; sequences without alignment specified will be
treated as if their alignment is the exact time the scheduler starts up
or detects that they have been modified.