add interruptible
This is the first of several commits laying the groundwork for full
support of running scheduled tasks. It introduces the interruptible
object, which is the basis for controlling continuously-actuated
peripherals.

An interruptible is a wrapper around a controller that sets the state
of the controlled peripheral according to a set of active interrupts.
The peripheral's state is set to the state of the highest-priority
interrupt, or the most recent in the case of a tie. It also exposes the
set of active interrupts and the current state of the peripheral.