Add timeSequenceExecutor
This is the first sequenceExecutor. SequenceExecutors extend
interruptibles and provide startSchedule, endSchedule, and
replaceSequenceItems methods.

On startSchedule, the timeSequenceExecutor starts controlling the
interruptible according to the sequenceItems. On replaceSequenceItems
the executor resets the schedule using the new sequenceItems. On
endSchedule, the executor removes any scheduled interrupts, leaving the
interruptible in the defaultState specified in the constructor.

The purpose of the sequenceExecutors is to handle the regular recurring
elements of the control. A higher level controller will start the
sequenceExecutors and keep track of the override rules. The combination
of interruptibles and layered control is intended to eventually support
triggered peripherals as well as continuous ones.