Remove alignment from duration sequences
When I made duration sequences, I wanted to do the right thing in the
case where the process was stopped and then restarted. If the sequence
was supposed to be off for an hour, then on for a minute, and was
stopped and restarted after 59 minutes, I wanted a mechanism to not
wait another hour after restarting. The alignment was meant to solve
that by remembering a start time for the sequence, but it wasn't very
good.

First, it added a lot of complexity to the durationSequenceExecutor.
The executor had to constantly figure out how far we were from the
alignment and where in the sequence we should be. Second, the alignment
was either implemented incorrectly or unpredictable--implemented as a
datetime, the sequence would drift unless it was evenly divisible by
the number of seconds in a day; implemented as a time of day, the
sequence would behave incorrectly when the alignment wrapped. Most
importantly, while using the program I found that the alignment was
confusing and never what I wanted.