delete sequences on peripheral delete.
I had an idea that a sequence might be separate from a peripheral, like
you could set up a 1 minute on / 1 minute off sequence and apply it to
multiple peripherals. But that's down the road, so the correct behavior
for now is to delete sequences and sequence items when peripherals are
deleted and not clutter up the db.