Make all dependencies injected.
I encountered a showstopper issue when trying to use the
request-adapter in a ui project built with webpack--when requiring
water-shape, webpack tried to package all of water-shape's
dependencies, including the hapi server and sqlite, as well as
additional copies of lodash and async.

The obvious way to address this was
to make water-shape require that client code inject its dependencies.
This is an unfortunate solution, because it makes the library much more
cumbersome to use and opens the door to tons of issues as the
dependencies move forward and water-shape can't be compatible with all
versions.

The only other solution I saw was to break water-shape into at least 3
different npm modules: one for dependency-injected modules, one for
generic functionality, and one for non-dependency injected modules.
Probably there would also have to be a fourth npm module for shared
test functionality for the other three. That approach seemed
potentially insane, so I went with 100% DI and hope to keep ahead of
the issues that will cause.